+++ TradeSkillMaster/ChangeLog.md
@@ -1,5 +1,7 @@
-## v4.7.10 Changes
+## v4.7.11 Changes
 
-* Auctioning and Crafting operation price value fixes.
+* Auctioning operation cancel repost threshold price value fixes.
+* Reverting the change that only allowed buying full stack sizes with Vendoring.
+* Some other smaller fixes.
 
 [Known Issues](http://support.tradeskillmaster.com/display/KB/TSM4+Currently+Known+Issues)